+++ web/src/app/theme.ts
@@ -1,70 +1,75 @@
-import { createStyles, createTheme } from '@material-ui/core/styles'
+import { createTheme, adaptV4Theme } from '@mui/material/styles'
+import createStyles from '@mui/styles/createStyles'
 // Theme documentation: https://material-ui.com/customization/palette/
 // Theme demo: https://material.io/resources/color/#!/?view.left=1&view.right=1&primary.color=003365&secondary.color=FBC02D
 // Do not export this directly for styling! theme should be accessed within makeStyles & withStyles. Use ErrorMessage.tsx as a reference
-export const theme = createTheme({
-  palette: {
-    primary: {
-      light: '#3E5C93',
-      main: '#003366',
-      dark: '#000C3A'
+export const theme = createTheme(
+  adaptV4Theme({
+    palette: {
+      primary: {
+        light: '#3E5C93',
+        main: '#003366',
+        dark: '#000C3A'
+      },
+      secondary: {
+        light: '#FFF263',
+        main: '#FBC02D',
+        dark: '#C49000'
+      },
+      success: { main: '#44D77A' },
+      error: { main: '#FF3E34' },
+      warning: { main: '#FE7921' },
+      contrastThreshold: 3,
+      tonalOffset: 0.1
     },
-    secondary: {
-      light: '#FFF263',
-      main: '#FBC02D',
-      dark: '#C49000'
+    typography: {
+      button: {
+        textTransform: 'none'
+      }
     },
-    success: { main: '#44D77A' },
-    error: { main: '#FF3E34' },
-    warning: { main: '#FE7921' },
-    contrastThreshold: 3,
-    tonalOffset: 0.1
-  },
-  typography: {
-    button: {
-      textTransform: 'none'
+    breakpoints: {
+      values: {
+        xs: 0,
+        sm: 600,
+        md: 1080, // Default: 960
+        lg: 1280,
+        xl: 1920
+      }
     }
-  },
-  breakpoints: {
-    values: {
-      xs: 0,
-      sm: 600,
-      md: 1080, // Default: 960
-      lg: 1280,
-      xl: 1920
-    }
-  }
-})
+  })
+)
 
-export const fireTableTheme = createTheme({
-  ...theme,
-  overrides: {
-    MuiTableCell: {
-      root: {
-        padding: 2
+export const fireTableTheme = createTheme(
+  adaptV4Theme({
+    ...theme,
+    overrides: {
+      MuiTableCell: {
+        root: {
+          padding: 2
+        },
+        head: {
+          fontWeight: 'bold',
+          minWidth: 30,
+          padding: 1,
+          paddingLeft: 7
+        },
+        stickyHeader: {
+          padding: 5
+        }
       },
-      head: {
-        fontWeight: 'bold',
-        minWidth: 30,
-        padding: 1,
-        paddingLeft: 7
+      MuiInputBase: {
+        root: {
+          fontSize: '1em'
+        }
       },
-      stickyHeader: {
-        padding: 5
-      }
-    },
-    MuiInputBase: {
-      root: {
-        fontSize: '1em'
-      }
-    },
-    MuiOutlinedInput: {
-      root: {
-        padding: 0
+      MuiOutlinedInput: {
+        root: {
+          padding: 0
+        }
       }
     }
-  }
-})
+  })
+)
 
 export const BACKGROUND_COLOR = { backgroundColor: '#e9ecf5' }