+++ web/src/features/fbaCalculator/components/FireCenterDropdown.tsx
@@ -1,5 +1,4 @@
-import { TextField } from '@material-ui/core'
-import { Autocomplete } from '@material-ui/lab'
+import { TextField, Autocomplete } from '@mui/material'
 import { FireCenter } from 'api/fbaAPI'
 import { isEqual } from 'lodash'
 import React from 'react'