+++ web/src/features/hfiCalculator/components/StaticCells.tsx
@@ -1,6 +1,5 @@
-import { TableCell } from '@material-ui/core'
-import { WeatherStation } from 'api/hfiCalcAPI'
-import { StationDaily } from 'api/hfiCalculatorAPI'
+import { TableCell } from '@mui/material'
+import { StationDaily, WeatherStation } from 'api/hfiCalculatorAPI'
 import HFICell from 'components/HFICell'
 import EmptyStaticCells from 'features/hfiCalculator/components/EmptyStaticCells'
 import HighestDailyFIGCell from 'features/hfiCalculator/components/HighestDailyFIGCell'