+++ CHANGELOG.md
@@ -1,8 +1,20 @@
-## 2022-04-19 HFI Calc [#1869](https://github.com/bcgov/wps/issues/1869)
+## 2022-04-26 HFI Calc [#1869](https://github.com/bcgov/wps/issues/1869)
 
 ### Features
 
 - **hfi calculator:** Introduce role permission for editing HFI prep fire starts.
+
+## 2022-04-21 HFI Calc
+
+### Bug
+
+- Fixed [#1904](https://github.com/bcgov/wps/issues/1904): Initial load of the page would not render loading spinner
+
+## 2022-04-19 HFI Calc
+
+### Features
+
+- **hfi calculator [#1849](https://github.com/bcgov/wps/issues/1849):** Added West Kelowna (1277) station with fuel type C7 to Penticton planning area as 2nd station in list. Added Station Bay 2 (1359) station with fuel type C5 to Vernon planning area as 3rd station in list
 
 ## 2022-04-14 HFI Calc - Bug Fix - PDF Download